rename workflow module to automation.
