+++ composer.json
@@ -68,7 +68,7 @@
             "Webkul\\User\\": "packages/Webkul/User/src",
             "Webkul\\Warehouse\\": "packages/Webkul/Warehouse/src",
             "Webkul\\WebForm\\": "packages/Webkul/WebForm/src",
-            "Webkul\\Workflow\\": "packages/Webkul/Workflow/src",
+            "Webkul\\Automation\\": "packages/Webkul/Automation/src",
             "Webkul\\DataGrid\\": "packages/Webkul/DataGrid/src"
         }
     },
