+++ config/app.php
@@ -244,7 +244,7 @@ return [
         Webkul\User\Providers\UserServiceProvider::class,
         Webkul\Warehouse\Providers\WarehouseServiceProvider::class,
         Webkul\WebForm\Providers\WebFormServiceProvider::class,
-        Webkul\Workflow\Providers\WorkflowServiceProvider::class,
+        Webkul\Automation\Providers\WorkflowServiceProvider::class,
     ],
 
     /*
