+++ config/concord.php
@@ -17,7 +17,7 @@ return [
         \Webkul\UI\Providers\ModuleServiceProvider::class,
         \Webkul\User\Providers\ModuleServiceProvider::class,
         \Webkul\WebForm\Providers\ModuleServiceProvider::class,
-        \Webkul\Workflow\Providers\ModuleServiceProvider::class,
+        \Webkul\Automation\Providers\ModuleServiceProvider::class,
     ],
     'register_route_models' => true,
 ];
