+++ packages/Webkul/Admin/src/Resources/views/settings/workflows/edit.blade.php
@@ -705,7 +705,7 @@
 
                 data() {
                     return {
-                        events: @json(app('\Webkul\Workflow\Helpers\Entity')->getEvents()),
+                        events: @json(app('\Webkul\Automation\Helpers\Entity')->getEvents()),
 
                         event: '{{ $workflow->event }}',
 
@@ -819,7 +819,7 @@
                 
                 data() {
                     return {
-                        conditions: @json(app('\Webkul\Workflow\Helpers\Entity')->getConditions()),
+                        conditions: @json(app('\Webkul\Automation\Helpers\Entity')->getConditions()),
 
                         conditionOperators: {
                             'price': [{
@@ -1046,7 +1046,7 @@
 
                 data() {
                     return {
-                        actions: @json(app('\Webkul\Workflow\Helpers\Entity')->getActions()),
+                        actions: @json(app('\Webkul\Automation\Helpers\Entity')->getActions()),
                     };
                 },
