+++ packages/Webkul/Automation/src/Contracts/Workflow.php
@@ -1,0 +1,5 @@
+<?php
+
+namespace Webkul\Automation\Contracts;
+
+interface Workflow {}
