+++ packages/Webkul/Automation/src/Helpers/Entity/Quote.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Webkul\Workflow\Helpers\Entity;
+namespace Webkul\Automation\Helpers\Entity;
 
 use Illuminate\Support\Arr;
 use Illuminate\Support\Facades\Http;
@@ -10,15 +10,16 @@ use Webkul\Attribute\Repositories\AttributeRepository;
 use Webkul\Contact\Repositories\PersonRepository;
 use Webkul\EmailTemplate\Repositories\EmailTemplateRepository;
 use Webkul\Lead\Repositories\LeadRepository;
+use Webkul\Quote\Repositories\QuoteRepository;
 
-class Person extends AbstractEntity
+class Quote extends AbstractEntity
 {
     /**
      * Define the entity type.
      *
      * @var string
      */
-    protected $entityType = 'persons';
+    protected $entityType = 'quotes';
 
     /**
      * Create a new repository instance.
@@ -28,6 +29,7 @@ class Person extends AbstractEntity
     public function __construct(
         protected AttributeRepository $attributeRepository,
         protected EmailTemplateRepository $emailTemplateRepository,
+        protected QuoteRepository $quoteRepository,
         protected LeadRepository $leadRepository,
         protected PersonRepository $personRepository
     ) {}
@@ -35,13 +37,13 @@ class Person extends AbstractEntity
     /**
      * Listing of the entities.
      *
-     * @param  \Webkul\Contact\Contracts\Person  $entity
-     * @return \Webkul\Contact\Contracts\Person
+     * @param  \Webkul\Quote\Contracts\Quote|int  $entity
+     * @return \Webkul\Quote\Contracts\Quote
      */
     public function getEntity($entity)
     {
-        if (! $entity instanceof \Webkul\Contact\Contracts\Person) {
-            $entity = $this->personRepository->find($entity);
+        if (! $entity instanceof \Webkul\Quote\Contracts\Quote) {
+            $entity = $this->quoteRepository->find($entity);
         }
 
         return $entity;
@@ -58,6 +60,10 @@ class Person extends AbstractEntity
 
         return [
             [
+                'id'         => 'update_quote',
+                'name'       => __('admin::app.settings.workflows.update-quote'),
+                'attributes' => $this->getAttributes('quotes'),
+            ], [
                 'id'         => 'update_person',
                 'name'       => __('admin::app.settings.workflows.update-person'),
                 'attributes' => $this->getAttributes('persons'),
@@ -68,6 +74,10 @@ class Person extends AbstractEntity
             ], [
                 'id'      => 'send_email_to_person',
                 'name'    => __('admin::app.settings.workflows.send-email-to-person'),
+                'options' => $emailTemplates,
+            ], [
+                'id'      => 'send_email_to_sales_owner',
+                'name'    => __('admin::app.settings.workflows.send-email-to-sales-owner'),
                 'options' => $emailTemplates,
             ], [
                 'id'              => 'trigger_webhook',
@@ -90,26 +100,32 @@ class Person extends AbstractEntity
     /**
      * Execute workflow actions.
      *
-     * @param  \Webkul\Workflow\Contracts\Workflow  $workflow
-     * @param  \Webkul\Contact\Contracts\Person  $person
+     * @param  \Webkul\Automation\Contracts\Workflow  $workflow
+     * @param  \Webkul\Quote\Contracts\Quote  $quote
      * @return array
      */
-    public function executeActions($workflow, $person)
+    public function executeActions($workflow, $quote)
     {
         foreach ($workflow->actions as $action) {
             switch ($action['id']) {
+                case 'update_quote':
+                    $this->quoteRepository->update([
+                        'entity_type'        => 'quotes',
+                        $action['attribute'] => $action['value'],
+                    ], $quote->id);
+
+                    break;
+
                 case 'update_person':
                     $this->personRepository->update([
                         'entity_type'        => 'persons',
                         $action['attribute'] => $action['value'],
-                    ], $person->id);
+                    ], $quote->person_id);
 
                     break;
 
                 case 'update_related_leads':
-                    $leads = $this->leadRepository->findByField('person_id', $person->id);
-
-                    foreach ($leads as $lead) {
+                    foreach ($quote->leads as $lead) {
                         $this->leadRepository->update([
                             'entity_type'        => 'leads',
                             $action['attribute'] => $action['value'],
@@ -127,9 +143,27 @@ class Person extends AbstractEntity
 
                     try {
                         Mail::queue(new Common([
-                            'to'      => data_get($person->emails, '*.value'),
-                            'subject' => $this->replacePlaceholders($person, $emailTemplate->subject),
-                            'body'    => $this->replacePlaceholders($person, $emailTemplate->content),
+                            'to'      => data_get($quote->person->emails, '*.value'),
+                            'subject' => $this->replacePlaceholders($quote, $emailTemplate->subject),
+                            'body'    => $this->replacePlaceholders($quote, $emailTemplate->content),
+                        ]));
+                    } catch (\Exception $e) {
+                    }
+
+                    break;
+
+                case 'send_email_to_sales_owner':
+                    $emailTemplate = $this->emailTemplateRepository->find($action['value']);
+
+                    if (! $emailTemplate) {
+                        break;
+                    }
+
+                    try {
+                        Mail::queue(new Common([
+                            'to'      => $quote->user->email,
+                            'subject' => $this->replacePlaceholders($quote, $emailTemplate->subject),
+                            'body'    => $this->replacePlaceholders($quote, $emailTemplate->content),
                         ]));
                     } catch (\Exception $e) {
                     }
@@ -141,7 +175,7 @@ class Person extends AbstractEntity
                         try {
                             $this->triggerWebhook(
                                 $action['hook'],
-                                $person
+                                $quote
                             );
                         } catch (\Exception $e) {
                             report($e);
@@ -157,10 +191,10 @@ class Person extends AbstractEntity
      * Trigger webhook.
      *
      * @param  array  $hook
-     * @param  \Webkul\Contact\Contracts\Person  $person
+     * @param  \Webkul\Quote\Contracts\Quote  $quote
      * @return void
      */
-    private function triggerWebhook($hook, $person)
+    private function triggerWebhook($hook, $quote)
     {
         if (in_array($hook['method'], ['get', 'delete'])) {
             Http::withHeaders(
@@ -173,7 +207,7 @@ class Person extends AbstractEntity
                 $this->formatHeaders($hook)
             )->{$hook['method']}(
                 $hook['url'],
-                $this->getRequestBody($hook, $person)
+                $this->getRequestBody($hook, $quote)
             );
         }
     }
@@ -203,14 +237,14 @@ class Person extends AbstractEntity
      * Prepare request body.
      *
      * @param  array  $hook
-     * @param  \Webkul\Contact\Contracts\Person  $person
+     * @param  \Webkul\Quote\Contracts\Quote  $quote
      * @return array
      */
-    private function getRequestBody($hook, $person)
+    private function getRequestBody($hook, $quote)
     {
-        $hook['simple'] = str_replace('person_', '', $hook['simple']);
+        $hook['simple'] = str_replace('quote_', '', $hook['simple']);
 
-        $results = $this->personRepository->find($person->id)->get($hook['simple'])->first()->toArray();
+        $results = $this->quoteRepository->find($quote->id)->get($hook['simple'])->first()->toArray();
 
         if (isset($hook['custom'])) {
             $customUnformatted = preg_split("/[\r\n,]+/", $hook['custom']);
