+++ packages/Webkul/Automation/src/Providers/ModuleServiceProvider.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace Webkul\Automation\Providers;
+
+use Webkul\Core\Providers\BaseModuleServiceProvider;
+use Webkul\Automation\Models\Workflow;
+
+class ModuleServiceProvider extends BaseModuleServiceProvider
+{
+    /**
+     * Define the modals to map with this module.
+     *
+     * @var array
+     */
+    protected $models = [
+        Workflow::class,
+    ];
+}
