--- packages/Webkul/Workflow/src/Contracts/Workflow.php
@@ -1,5 +1,0 @@
-<?php
-
-namespace Webkul\Workflow\Contracts;
-
-interface Workflow {}
